requirements.txt
anthropic>=0.49.0
feedparser>=6.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.28.0
chardet>=5.0.0
jinja2>=3.1.0
